Split expression tests into their own class
