+++ src/XMakeBuildEngine/UnitTests/Microsoft.Build.Engine.UnitTests.csproj
@@ -147,6 +147,7 @@
     <Compile Include="Evaluation\ProjectRootElementCache_Tests.cs" />
     <Compile Include="Evaluation\ProjectStringCache_Tests.cs" />
     <Compile Include="EventArgsFormatting_Tests.cs" />
+    <Compile Include="ExpressionTreeExpression_Tests.cs" />
     <Compile Include="ExpressionTree_Tests.cs" />
     <Compile Include="FileLogger_Tests.cs" />
     <Compile Include="HashTableUtility_Tests.cs" />
